requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
pydantic-settings==2.1.0
structlog==24.1.0
python-multipart==0.0.6
argon2-cffi==23.1.0
pyjwt==2.8.0
markitdown==0.1.5
markdown-it-py==3.0.0
jieba==0.42.1
pgvector==0.2.4
sse-starlette==2.0.0
python-docx==1.1.0
pytest==8.0.0
pytest-asyncio==0.23.5
